Fix bugs

+++ src/screen.cpp
@@ -5,13 +5,12 @@ Screen::Screen(size_t width, size_t height) : width_(width), height_(height) {
     data_.reserve(width * height);
 }
 
-DiscreteVector2
-Screen::projectVertexToScreenAndDiscretize(const Primitives::Vertex &v) const {
+DiscreteVector2 Screen::projectVertexToScreenAndDiscretize(const Primitives::Vertex &v) const {
     auto cutToScreen = [](double value, const size_t bound) {
-        return std::max(0.0, std::min(value, bound - 1.0));
+        return size_t(std::max(0.0, std::min(value, bound - 1.0)));
     };
-    double x = cutToScreen((v.getPosition().x + 1) / 2 * width_, width_);
-    double y = cutToScreen((v.getPosition().y + 1) / 2 * height_, height_);
+    size_t x = cutToScreen((v.getPosition().x + 1) / 2 * width_, width_);
+    size_t y = cutToScreen((v.getPosition().y + 1) / 2 * height_, height_);
     return sf::Vector2i{x, y};
 }
 
@@ -21,9 +20,7 @@ void Screen::setPixel(size_t i, size_t j, const Color &color) {
     }
 }
 
-bool Screen::isPixelValid(size_t i, size_t j) const {
-    return 0 <= i && i < height_ && 0 <= j && j < width_;
-}
+bool Screen::isPixelValid(size_t i, size_t j) const { return 0 <= i && i < height_ && 0 <= j && j < width_; }
 
 const std::vector<sf::Vertex> &Screen::getDrawData() const { return data_; }
